LayoutTests/platform/ios-simulator/editing/selection/move-by-word-visually-multi-line-expected.txt

CONSOLE MESSAGE: line 16: TypeError: null is not an object (evaluating 'positions[i].node.tagName')
אאא אאא
abc def ghi jkl mn 


opq rst uvw xyz
abc def ghi jkl mn



opq rst uvw xyz
אאא אאא אאא אאא 


אאא אאא אאא אאא
אאא אאא אאא אאא 



אאא אאא אאא אאא
abc def אאא אאא hij אאא אאא uvw xyz 


אאא kj אאא mn opq אאא אאא
abc def אאא אאא hij אאא אאא uvw xyz 



אאא kj אאא mn opq אאא אאא
abc def ghi jkl mn 


opq rst uvw xyz
abc def ghi jkl mn



opq rst uvw xyz
אאא אאא אאא אאא 


אאא אאא אאא אאא
אאא אאא אאא אאא



אאא אאא אאא אאא
abc def אאא אאא hij אאא אאא uvw xyz 


אאא kj אאא mn opq אאא אאא
abc def אאא אאא hij אאא אאא uvw xyz



אאא kj אאא mn opq אאא אאא
abc def ghi jkl mn
opq rst uvw xyz
abc def ghi jkl mn
abc def ghi jkl mn

opq rst uvw xyz
abc def inside span hij opq
00

======== Move By Word ====
Test 1, LTR:
Move right by one word
"abc def ghi jkl mn "[0, 4, 8, 12, 16], "opq rst uvw xyz"[0, 4, 8, 12, 15]
Move left by one word
"opq rst uvw xyz"[15, 12, 8, 4, 0], "abc def ghi jkl mn "[16, 12, 8, 4, 0]
Test 2, LTR:
Move right by one word
"abc def ghi jkl mn "[0, 4, 8, 12, 16], "opq rst uvw xyz"[0, 4, 8, 12, 15]
Move left by one word
"opq rst uvw xyz"[15, 12, 8, 4, 0], "abc def ghi jkl mn "[16, 12, 8, 4, 0]
Test 3, LTR:
Move right by one word
"AAA AAA AAA AAA "[0, 3, 8, 11], "AAA AAA AAA AAA"[0, 3, 8, 11, 15]
Move left by one word
"AAA AAA AAA AAA"[15, 11, 8, 3, 0], "AAA AAA AAA AAA "[11, 8, 3, 0]
Test 4, LTR:
Move right by one word
"AAA AAA AAA AAA "[0, 3, 8, 11], "AAA AAA AAA AAA"[0, 3, 8, 11, 15]
Move left by one word
"AAA AAA AAA AAA"[15, 11, 8, 3, 0], "AAA AAA AAA AAA "[11, 8, 3, 0]
Test 5, LTR:
Move right by one word
" abc def AAA AAA hij AAA AAA uvw xyz "[1, 5, 9, 12, 17, 21, 25, 29, 33], "AAA kj AAA mn opq AAA AAA"[0, 4, 7, 11, 14, 18, 22, 25]
Move left by one word
"AAA kj AAA mn opq AAA AAA"[25, 22, 18, 14, 11, 7, 4, 0], " abc def AAA AAA hij AAA AAA uvw xyz "[33, 29, 25, 21, 17, 12, 9, 5, 1]
Test 6, LTR:
Move right by one word
